add profile handling and refactor code
