+++ src/components/profile-configuration/ProfileConfiguration.module.scss
@@ -3,9 +3,7 @@
     margin-top: 1rem;
   }
   .pills {
-    margin: 1rem 0;
     .selected {
-      margin: 1rem 0;
     }
     .options {
     }
