+++ src/components/profile-configuration/index.js
@@ -1,0 +1,1 @@
+export * from './ProfileConfiguration';
